Fix text wrap && backspace deleting 0
Fix a bug when user delete using backspace if display shows 0 the
display would shrink as no number is in to fill it. Now this behavior is
removed.
Fix When user would write a large number of numbers , the numbers would
then escape the boundaries of the calculator. Now this behavior is fixed